+++ docs/source/changelog.rst
@@ -21,7 +21,9 @@ Version 2.0.0
 
     * :func:`~.get_names` now returns :class:`tuple` instead of :class:`list`.
     * Bumped Lightning version to ``>=2.5.0`` (:issue:`29`).
-    * :func:`~.upsample_pcd` moved from :mod:`~aidsorb.data` to :mod:`~aidsorb.transforms`.
+    * :func:`~.upsample_pcd` moved from :mod:`~aidsorb.data` to
+      :mod:`~aidsorb.transforms` and now raises error if target size is not
+      greater than the original.
     * :func:`~.split_pcd` moved from :mod:`~aidsorb.utils` to
       :mod:`~aidsorb.transforms` and *no longer copies data*.
     * :mod:`~.transforms` use :mod:`torch` instead of :mod:`numpy` (:issue:`32`).